requirements.txt
django-environ==0.4.5
django-filter==1.1.0
django-graphql-jwt==0.1.5
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.1
graphql-relay==0.4.5
iso8601==0.1.12
promise==2.2.1
PyJWT==1.6.4
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
